+++ email_auth/conftest.py
@@ -23,3 +23,12 @@ def mock_email_verification_qs():
         "email_auth.models.EmailVerification.objects", new=mock_qs
     ):
         yield mock_qs
+
+
+@pytest.fixture
+def mock_password_reset_qs():
+    mock_qs = mock.Mock(spec=models.PasswordReset.objects)
+    mock_qs.all.return_value = mock_qs
+
+    with mock.patch("email_auth.models.PasswordReset.objects", new=mock_qs):
+        yield mock_qs